Version 1 (#68)
* explode model

* exploding tests

* ExplodingModel tests

* tests refactor

* explode relationships

* refactor for exploding relationships

* start testing relationships

* container and software system relationships

* component relationships

* infrastructure nodes relationships

* relationships exploded

* exploding relationships finished

* explode views mappings

* Explode views tests first 4

* Explode Views

* exploding documentation

* doc elements explode tests

* exploding doc tests

* Elements detector

* basic tests sceleton

* more tests

* people and instances tests

* some clean

* basic elements detector tests

* rise events on element data changed

* start relationship detector

* detect add relationship  tests

* relationships detector tests

* publish relationship changed event

* start view detection

* first working test

* Detect View element tests

* detect view ids test

* detect static diagrams

* Detect deployment views

* detect image view

* detect dynamic views

* Filtered View Detect

* Start Import command

* startimport command tests

* ImportElementsCommand

* Test class

* Import elements command tests

* import relationships

* import relationships tests

* import views command

* save workspace views information

* import views tests

* clear docs and images

* import documentation

* use base import data handler

* import documentation tests

* import image

* save decision changes

* decsion mapping

* finalize import

* start import endpoint

* update .net deps

* import endpoints

* add vscode to git ignore

* fix import issues

* import view config

* clear decisions

* docueye api clinet

* Start cli changes

* cli work in progress

* import process in cli

* import success

* fix decision import

* fix elements diagram import

* remove old importer code

* clear old documentation save actions

* fix clear old documentation import

* import flow in cli

* fix existing tests

* import clear doc items tests

* import decisions tests

* remove cli client tests

* update local buid scripts

* documentation changes

* sln changes

* move code

* new json model

* New cli (#62)

* command line refactor

* working on console error reporting

* first import

* run localy

* delete workspace

* working on export to json

* working on deployment view import

* working on serialization

* fix context

* deployment view fixed

* fix dsl identifier

* working import for example

* fix tests

* export command

* cli descriptions

* sln cleanup

* default theme fix

* tags fix

* instanceId fix

* set identifiers at workspace level

* fix documentation

* fix docueye properties

* version check

* version check and logging

* log formatter

* Deps update (#63)

* update .net 8 deps and fix tests

* rc5

* remove automapper (#64)

* exploder maps no views

* relationship in view map

* all exploder mappings

* refactor cli mapping

* remove old exploder mappings

* assert helpers separate

* add projects for api model maps

* api to model mapping

* modelkeeper maps

* viewskeeper maps

* docskeeper maps

* all maps

* start removing automapper

* remove automapper

* remove mediatr (#65)

* add mediator

* remove mediatr

* mediator service collection

* use assemblies

* run app

* cleanup

* fix void command no implementation

* bump to .net 10 (#66)

* update images versions

* Missing commands (#67)

* import openapi file

* fix openapiimport

* commands

* delete openapi

* docs workspaces

* cli docs

* cli docs

* cli docs

* docs

* cli readme

* docs and scripts

* fix readme

* fixes in dsl parsing

* dsl parsing fixes

* fix import workspace configuration

* fix username mapping

* show validation errors in import response

* compose update

* update mongo to 8.2.2

* rm old command line parser

* fix cli

* clean up

* fix import script

* fix dotnet versions in pipelines

+++ Documentation/import-workspace.ps1
@@ -9,21 +9,29 @@ Param(
     [string]$importKey = (New-Guid).Guid,
     [Parameter(Mandatory,
     HelpMessage="DocuEye address ex. http://localhost:8080")]
-    [string]$docueyeAddress
+    [string]$docueyeAddress,
+    [Parameter(
+    HelpMessage="Indicates that script will use dotnet tool instead of docueye-cli docker image")]
+    [switch]$useDotNetTool
 )
 
 ####
-# .\import-workspace.ps1 -docueyeAddress http://localhost:8080 -adminToken docueyedmintoken -workspaceId 8ff549b9-af3b-4e76-9cdf-aaa3218398a4 
+# .\import.ps1 -docueyeAddress http://localhost:8080 -adminToken docueyedmintoken -workspaceId 638d0822-12c7-4998-8647-9c7af7ad2989 
 ####
 
-
-#export workspace to json format
-docker run -it --rm -v "$($PWD):/usr/local/structurizr" structurizr/cli export --workspace workspace.dsl -format json
-#import workspace to DocuEye
+if(!$useDotNetTool) {
 docker run -it --rm --network="host" -v "$($PWD):/app/import" jacekzwpl/docueye-cli  `
---import=workspace  `
---docueyeAddress="$docueyeAddress"  `
---adminToken="$adminToken"  `
---importKey="$importKey"  `
---workspaceId="$workspaceId"  `
---workspaceFile=./import/workspace.json
+workspace import  `
+--docueye-address="$docueyeAddress"  `
+--admin-token="$adminToken"  `
+--key="$importKey"  `
+--id="$workspaceId"  `
+--file=./import/workspace.dsl
+}else {
+docueye workspace import  `
+--docueye-address="$docueyeAddress"  `
+--admin-token="$adminToken"  `
+--key="$importKey"  `
+--id="$workspaceId"  `
+--file=workspace.dsl
+}
